requirements.txt
pandas==1.5.3
numpy==1.24.3
psycopg2-binary==2.9.7
boto3==1.26.137
